Reverting back to Parameters update route works new jar
