+++ src/main/java/com/schindig/utils/UserParams.java
@@ -1,0 +1,46 @@
+package com.schindig.utils;
+/**
+ * Created by Agronis on 12/14/15.
+ */
+public class UserParams {
+
+    /**
+     * In relation to User items
+     **/
+    public Integer userId;
+    public String username;
+    public String phone;
+    public String email;
+    public String firstName;
+    public String lastName;
+    public String rsvpStatus;
+
+    public Integer getUserId() {
+
+        return userId;
+    }
+    public String getUsername() {
+
+        return username;
+    }
+    public String getPhone() {
+
+        return phone;
+    }
+    public String getEmail() {
+
+        return email;
+    }
+    public String getFirstName() {
+
+        return firstName;
+    }
+    public String getLastName() {
+
+        return lastName;
+    }
+    public String getRsvpStatus() {
+
+        return rsvpStatus;
+    }
+}
